Рефакторит код, разбивает презентер на дополнительные модули
