+++ src/model/destinations-model.js
@@ -1,0 +1,7 @@
+import { getDestinations } from '../fish/destinations.js';
+
+export default class DestinationsModel {
+  get() {
+    return getDestinations();
+  }
+}
