+++ src/utils.js
@@ -1,6 +1,6 @@
 import dayjs from 'dayjs';
 
-const getRandomInteger = (from, to) => {
+export const getRandomInteger = (from, to) => {
   if (to <= from) {
     throw new Error('Начальная точка диапазона должна быть меньше конечной точки диапазона');
   }
@@ -8,14 +8,14 @@ const getRandomInteger = (from, to) => {
   return Math.floor(Math.random() * (to - from + 1)) + from;
 };
 
-const getRandomArrayElement = (elements) => {
+export const getRandomArrayElement = (elements) => {
   const maxIndex = elements.length - 1;
   const randomIndex = getRandomInteger(0, maxIndex);
 
   return elements[randomIndex];
 };
 
-const createRandomizerOfUniqueInteger = (from, to) => {
+export const createRandomizerOfUniqueInteger = (from, to) => {
   const repeatedNumbers = [];
 
   return () => {
@@ -31,44 +31,27 @@ const createRandomizerOfUniqueInteger = (from, to) => {
   };
 };
 
-const createCounter = () => {
+export const createCounter = () => {
   let count = 0;
 
   return () => ++count;
 };
 
-const capitalizeFirstLetter = (text) => {
+export const capitalizeFirstLetter = (text) => {
   const firstLetter = text[0].toUpperCase();
   const restText = text.slice(1);
 
   return `${firstLetter}${restText}`;
 };
 
-const getOffersByType = (offerGroups, type) => {
+export const getOffersByType = (offerGroups, type) => {
   const typeOffer = offerGroups.find((offerGroup) => (offerGroup.type === type));
 
   return typeOffer.offers;
 };
 
-const getOffersByIds = (offers, ids) => offers.filter((offer) => ids.includes(offer.id));
+export const getOffersByIds = (offers, ids) => offers.filter((offer) => ids.includes(offer.id));
 
-const getDestinationById = (destinations, id) => destinations.find((destination) => (destination.id === id));
+export const getDestinationById = (destinations, id) => destinations.find((destination) => (destination.id === id));
 
-const humanizeTime = (time) => dayjs(time).format('HH:mm');
-
-const machinizeTime = (time) => dayjs(time).format('YYYY-MM-[DD]T[HH]:mm');
-
-export const humanizeDate = (date) => dayjs(date).format('DD/MM/YY');
-
-export {
-  getRandomInteger,
-  getRandomArrayElement,
-  createRandomizerOfUniqueInteger,
-  createCounter,
-  capitalizeFirstLetter,
-  getOffersByType,
-  getOffersByIds,
-  getDestinationById,
-  humanizeTime,
-  machinizeTime,
-};
+export const formatDate = (date, format) => dayjs(date).format(format);
