+++ src/view/point-view.js
@@ -32,9 +32,8 @@ export default class PointView extends BaseView {
    */
   setIcon(name) {
     const element = this.querySelector('.event__type-icon');
-    const url = `img/icons/${name}.png`;
 
-    element.src = url;
+    element.src = `img/icons/${name}.png`;
 
     return this;
   }
